+++ src/main/java/com/yas/backend/domain/invitation/controller/InvitationCommandController.java
@@ -1,9 +1,9 @@
 package com.yas.backend.domain.invitation.controller;
 
-import com.yas.backend.domain.invitation.data.mapper.InvitationMapper;
+import com.yas.backend.domain.invitation.mapper.InvitationMapper;
 import com.yas.backend.domain.invitation.exchange.InvitationAnswerRequest;
 import com.yas.backend.domain.invitation.exchange.InvitationCreateRequest;
-import com.yas.backend.domain.invitation.service.InvitationCommandService;
+import com.yas.backend.domain.invitation.service.applicationservice.InvitationCommandService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -20,13 +20,13 @@ public class InvitationCommandController {
         invitationCommandService.create(invitationRequest.toDto());
     }
     
-    @PatchMapping(value = "/team/invitaion/{invitationId}/status")
+    @PatchMapping(value = "/invitaion/{invitationId}/status")
     public void answerJoin(@RequestBody final InvitationAnswerRequest request, @PathVariable("invitationId") Long invitationId) {
         invitationCommandService.updateStatus(request.toDto(invitationId));
     }
 
-    @DeleteMapping(value = "/team/invitation/{invitationId}/status")
+    @DeleteMapping(value = "/invitation/{invitationId}")
     public void deleteInvitation(@PathVariable("invitationId") Long invitationId) {
-        invitationCommandService.remove(invitationId);
+        invitationCommandService.updateStatus(invitationId);
     }
 }
