+++ src/main/java/com/yas/backend/domain/invitation/service/applicationservice/InvitationCommandService.java
@@ -1,19 +1,17 @@
-package com.yas.backend.domain.invitation.service;
+package com.yas.backend.domain.invitation.service.applicationservice;
 
-import com.yas.backend.common.enums.InvitationStatus;
+import com.yas.backend.common.enums.ErrorType;
 import com.yas.backend.common.exception.InvitationAlreadyExistException;
-import com.yas.backend.common.exception.TeamNotFoundException;
-import com.yas.backend.common.exception.UserNotFoundException;
+import com.yas.backend.common.exception.YasDomainValidationException;
 import com.yas.backend.domain.invitation.Invitation;
-import com.yas.backend.domain.invitation.data.mapper.InvitationMapper;
+import com.yas.backend.domain.invitation.mapper.InvitationMapper;
 import com.yas.backend.domain.invitation.dto.InvitationDto;
-import com.yas.backend.domain.invitation.service.domainService.InvitationService;
+import com.yas.backend.domain.invitation.service.InvitationService;
 import com.yas.backend.domain.user.mapper.UserMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import javax.validation.ValidationException;
 import java.time.LocalDateTime;
 
 @Slf4j
@@ -31,14 +29,11 @@ public class InvitationCommandService {
                 throw new InvitationAlreadyExistException();
             }
             Invitation invitation = Invitation.create(invitationDto);
-            invitation.setStatus(InvitationStatus.REQUEST);
-            invitation.removeInvitation(true);
-            invitation.setCreatedAt(LocalDateTime.now());
             return invitationService.save(invitationMapper.domainToDto(invitation)).getId();
 
-        } catch (InvitationAlreadyExistException | UserNotFoundException | TeamNotFoundException e) {
+        } catch (InvitationAlreadyExistException e) {
             log.info(e.getMessage());
-            throw new ValidationException();
+            throw new YasDomainValidationException(ErrorType.JOIN_ALREADY_EXIST.getMessage());
         }
     }
 
@@ -47,14 +42,5 @@ public class InvitationCommandService {
         invitation.setStatus(invitationDto.getStatus());
         invitation.setUpdatedAt(LocalDateTime.now());
         invitationService.save(invitationMapper.domainToDto(invitation));
-    }
-
-    public void remove(Long invitationId) {
-        Invitation invitation = Invitation.create(invitationService.findById(invitationId));
-
-        invitation.removeInvitation(false);
-        invitation.setUpdatedAt(LocalDateTime.now());
-        invitationService.save(invitationMapper.domainToDto(invitation));
-
     }
 }
