+++ src/main/java/com/yas/backend/domain/invitation/service/applicationservice/InvitationSearchService.java
@@ -1,11 +1,11 @@
-package com.yas.backend.domain.invitation.service;
+package com.yas.backend.domain.invitation.service.applicationservice;
 
 import com.querydsl.core.BooleanBuilder;
-import com.yas.backend.domain.invitation.data.mapper.InvitationMapper;
+import com.yas.backend.domain.invitation.mapper.InvitationMapper;
 import com.yas.backend.domain.invitation.dto.InvitationDto;
 import com.yas.backend.domain.invitation.dto.InvitationPredicate;
 import com.yas.backend.domain.invitation.exchange.InvitationSearchRequest;
-import com.yas.backend.domain.invitation.service.domainService.InvitationService;
+import com.yas.backend.domain.invitation.service.InvitationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
